Work around float limitation in CommandLineParser
Float data types cannot be declared as options, otherwise get an error
during parsing about invalid type.  Need to keep these delcared as
doubles.

Consequently, before using these options, want to cast them to float to
avoid extra casting later at every usage.

+++ src/Program.cs
@@ -91,11 +91,11 @@ namespace ScaffoldingGenerator
                     }
                     Console.WriteLine("Made support normals");
                     foreach (Vector3 supportNormal in supportNormals) {
-                        scaffoldingFacets.AddRange(GenerateLineScaffolding(model, largeRegions, supportNormal, opts.SupportSpacing, opts.PlateSpacing));
+                        scaffoldingFacets.AddRange(GenerateLineScaffolding(model, largeRegions, supportNormal, (float)opts.SupportSpacing, (float)opts.PlateSpacing));
                     }
                 }
                 if (opts.DoContourScaffolding) {
-                    scaffoldingFacets.AddRange(GenerateContourScaffolding(largeRegions, opts.PlateSpacing, edgeFacetIndex));
+                    scaffoldingFacets.AddRange(GenerateContourScaffolding(largeRegions, (float)opts.PlateSpacing, edgeFacetIndex));
                 }
                 StlBinaryWriter writer = new StlBinaryWriter();
                 writer.Write("out.stl", scaffoldingFacets.ToArray());
